foray_ui/nodes/pyproject.toml
[project]
name = "foray_nodes"
version = "0.0.1"
requires-python = ">=3.10"

dependencies = [
    "foray",
    "numpy >= 2.0",
    "pillow>=11.2.1",
]

[tool.pyright]
venv = ".venv"
venvPath = "./"

[tool.uv.sources]
foray = { path = "../foray_python_lib/" }
